chore: clean up unused code. use direct rbtc deposit on transak
