+++ src/app/components/TransakDialog/FiatButton.tsx
@@ -2,17 +2,22 @@ import { Button } from '@blueprintjs/core';
 import React from 'react';
 import { translations } from 'locales/i18n';
 import { Trans } from 'react-i18next';
-import { AssetRenderer } from '../../../components/AssetRenderer';
-import { Asset } from '../../../../types/asset';
+import { Asset } from 'types/asset';
+import { AssetRenderer } from '../AssetRenderer';
 
-interface Props {
+type FiatButtonProps = {
   loading: boolean;
   ready: boolean;
   onClick: () => void;
   disabled?: boolean;
-}
+};
 
-export function FiatButton({ loading, ready, onClick, disabled }: Props) {
+export const FiatButton: React.FC<FiatButtonProps> = ({
+  loading,
+  ready,
+  onClick,
+  disabled,
+}) => {
   return (
     <>
       <div className="tw-justify-center tw-items-center tw-flex tw-flex-row tw-w-full">
@@ -21,7 +26,7 @@ export function FiatButton({ loading, ready, onClick, disabled }: Props) {
           className="button-round tw-text-primary tw-text-lg tw-mx-auto"
           text={
             <Trans
-              i18nKey={translations.fastBtcDialog.fiatDialog.title}
+              i18nKey={translations.transakDialog.cta}
               components={[<AssetRenderer asset={Asset.RBTC} />]}
             />
           }
@@ -32,4 +37,4 @@ export function FiatButton({ loading, ready, onClick, disabled }: Props) {
       </div>
     </>
   );
-}
+};
