--- src/app/containers/FastBtcDialog/components/transaction.module.scss
@@ -1,30 +1,0 @@
-.title {
-  @apply tw-text-center tw-text-3xl tw-leading-relaxed tw-font-medium tw-normal-case;
-  margin-bottom: 57px;
-}
-
-.subtitle {
-  @apply tw-text-center tw-text-xl tw-leading-tight tw-font-light tw-normal-case;
-  margin-bottom: 27px;
-}
-
-.amountBlock {
-  @apply tw-inline-block tw-mx-auto tw-mt-0 tw-mb-9;
-}
-
-.amountBlockBtc {
-  @apply tw-text-lg tw-leading-tight tw-font-medium;
-}
-
-.amountBlockUsd {
-  @apply tw-text-sm tw-leading-snug tw-font-light;
-}
-
-.txData {
-  width: 230px;
-}
-
-.arrowDown {
-  width: 30px;
-  margin: 25px auto;
-}
