--- src/app/containers/FastBtcForm/index.module.scss
@@ -1,31 +1,0 @@
-.top-up-container {
-  @media screen and (max-width: 767px) {
-    & :global(.bp3-dialog) {
-      @apply tw-w-full tw-h-screen tw-m-0
-        tw-border-0 tw-rounded-none;
-    }
-  }
-
-  @screen lg {
-    & :global(.bp3-dialog) {
-      transition: width 0.5s;
-      will-change: width;
-      &.top-up-dialog-extend {
-        width: 800px;
-      }
-    }
-  }
-
-  .circle {
-    @apply tw-rounded-full;
-    width: 15px;
-    height: 15px;
-    
-    &.circle-red {
-      @apply tw-bg-warning;
-    }
-    &.circle-green {
-      @apply tw-bg-success;
-    }
-  }
-}
