+++ src/app/containers/WalletProvider/index.tsx
@@ -26,12 +26,6 @@ import { reducer, sliceKey } from './slice';
 import { walletProviderSaga } from './saga';
 import { selectRequestDialogState } from '../../../store/global/transactions-store/selectors';
 import { TxRequestDialog } from './components/TxRequestDialog';
-import { FastBtcForm } from '../FastBtcForm/Loadable';
-import {
-  sliceKey as btcSlice,
-  reducer as btcReducer,
-} from '../FastBtcForm/slice';
-import { fastBtcFormSaga } from '../FastBtcForm/saga';
 import { currentChainId } from '../../../utils/classifiers';
 import { actions } from './slice';
 import { useEvent } from 'app/hooks/useAnalytics';
@@ -51,9 +45,6 @@ export function WalletProvider(props: Props) {
 
   useInjectReducer({ key: transactionsSlice, reducer: transactionsReducer });
   useInjectSaga({ key: transactionsSlice, saga: transactionsStateSaga });
-
-  useInjectReducer({ key: btcSlice, reducer: btcReducer });
-  useInjectSaga({ key: btcSlice, saga: fastBtcFormSaga });
 
   const requestDialog = useSelector(selectRequestDialogState);
   const { bridgeChainId } = useSelector(selectWalletProvider);
@@ -81,7 +72,6 @@ export function WalletProvider(props: Props) {
       <WalletWatcher />
       <>{props.children}</>
       <TxRequestDialog {...requestDialog} />
-      <FastBtcForm />
     </SovrynWallet>
   );
 }
