+++ src/app/pages/BuySovPage/components/TopUpWallet/index.tsx
@@ -4,11 +4,11 @@ import { Button } from '../Button';
 import { useIsConnected } from '../../../../hooks/useAccount';
 import { useAssetBalanceOf } from '../../../../hooks/useAssetBalanceOf';
 import { Asset } from '../../../../../types/asset';
-import { FastBtcDialog } from '../../../../containers/FastBtcDialog';
 import { useTranslation } from 'react-i18next';
 import { translations } from '../../../../../locales/i18n';
 import { Trans } from 'react-i18next';
 import { AssetRenderer } from '../../../../components/AssetRenderer';
+import { TransackDialog } from 'app/components/TransakDialog/TransakDialog';
 
 export function TopUpWallet() {
   const { t } = useTranslation();
@@ -49,7 +49,7 @@ export function TopUpWallet() {
         disabled={!connected}
         onClick={() => setOpen(true)}
       />
-      <FastBtcDialog isOpen={open} onClose={() => setOpen(false)} />
+      <TransackDialog isOpen={open} onClose={() => setOpen(false)} />
     </Card>
   );
 }
