+++ src/locales/fr/translation.json
@@ -208,98 +208,6 @@
     "startMargin": "Marge de départ",
     "startPrice": "Prix ​​de départ"
   },
-  "fastBTC": {
-    "screens": {
-      "1": {
-        "createAddress": "Créer une adresse",
-        "p1": "Pour emprunter / prêter et trade sur Sovryn, vous devez d'abord activer votre Bitcoin en le transférant dans votre portefeuille rsk.",
-        "p2": "Pour ce faire, créez une adresse btc unique qui se rattache à votre adresse RSK, en suralimentant votre BTC en RBTC.",
-        "title": "Tu es connecté!"
-      },
-      "2": {
-        "contactUs": "Pour obtenir de l'aide, rejoignez-nous sur",
-        "depositRequirements": "Conditions de dépôt",
-        "fee": "Frais",
-        "feeExplainer": "Déduit du montant de la transaction",
-        "important": "Important",
-        "important1": "Ne déposez pas d'actifs non BTC à cette adresse, sinon vos actifs seront perdus définitivement",
-        "important2": "Veuillez prévoir jusqu'à 15 minutes pour que la transaction soit traitée",
-        "p1": "Votre adresse BTC unique a été créée, transférez du BTC pour recevoir du RBTC dans votre portefeuille RSK.",
-        "txDetected": "Transaction détectée"
-      },
-      "3": {
-        "amount": "Montant traité",
-        "estimated": "Temps estimé",
-        "fee": "Frais",
-        "from": "Du portefeuille",
-        "initiated": "Initié",
-        "status": "Status",
-        "to": "Vers le portefeuille",
-        "txDetails": "Détails de la transaction",
-        "txHash": "rBTC Tx hash",
-        "txId": "BTC Tx ID"
-      }
-    }
-  },
-  "fastBtcDialog": {
-    "buttonBTC": "Obtenez du rBTC avec du BTC",
-    "buttonFiat": "Obtenez du rBTC avec du Fiat",
-    "note": "Assurez-vous toujours de vérifier l'adresse de dépôt BTC lorsque FastBTC est mis à jour",
-    "fiatDialog": {
-      "explanation": "En cliquant sur le bouton ci-dessous, vous ouvrirez un fiat sur la rampe. Assurez-vous que l'adresse BTC à laquelle vous envoyez vos fonds est la même que celle ci-dessous. FastBTC effectuera la conversion en rBTC pour vous en arrière-plan. Gardez à l'esprit la limite de {{amount}}",
-      "title": "Achetez du rBTC avec du Fiat:"
-    },
-    "instructions": {
-      "point1": "Ne déposez rien d'autre que du BTC",
-      "point2": "N'envoyez pas plus de BTC que la limite maximum",
-      "point3": "Attendez 15 minutes pour que la transaction soit traitée",
-      "title": "Instructions:"
-    },
-    "limits": {
-      "btc": "BTC",
-      "max": "MAX: {{amount}}",
-      "min": "MIN: {{amount}}",
-      "title": "Limites de dépôt:"
-    },
-    "qr": {
-      "title": "Envoyez les BTC à cette adresse:"
-    },
-    "subtitle": "Déposez des BTC ou des fiat pour recevoir des rBTC dans votre portefeuille",
-    "title": "Obtenez du rBTC",
-    "transaction": {
-      "address": "Addresse",
-      "close": "Fermer",
-      "confirmed": "confirmé",
-      "fee": "Frais de transactions:",
-      "from": "De:",
-      "hash": "Hash de transaction:",
-      "pending": "en Attente ",
-      "sent": "envoyé",
-      "status": "Statut {{status}}",
-      "title": "Détails de la transaction",
-      "to": "À:"
-    }
-  },
-  "fastBtcForm": {
-    "connect": "Connectez-vous à votre portefeuille ou entrez simplement l'adresse du portefeuille au-dessus de vous-même.",
-    "connectWallet": "Pour utiliser Top-Up, vous devez connecter votre portefeuille avec dapp.",
-    "depositAddressCopied": "Votre adresse de dépôt BTC copiée dans le presse-papiers",
-    "depositAdress": "Adresse de dépôt BTC",
-    "disabledText": "Les dépôts complémentaires ont été temporairement désactivés. Nous vous demandons de ne pas envoyer de fonds avant d'être réactivé. Merci!",
-    "exchanging": "Nous échangeons du rBTC dès maintenant.",
-    "history": {
-      "empty": "L'historique est vide.",
-      "loading": "Chargement...",
-      "title": "Historique des dépôts BTC",
-      "walletHistory": "Entrez votre adresse de portefeuille RSK pour voir l'historique."
-    },
-    "placeholder": "Votre portefeuille RSK",
-    "swapAgain": "Échangez à nouveau",
-    "title": "Recharger",
-    "toDepositWallet": "vers le portefeuille de dépôt ",
-    "transfered": "a été transféré sur votre walllet RSK",
-    "waitingForDeposit": "En attente de dépôt"
-  },
   "fastBtcPage": {
     "meta": {
       "description": "Déplacez vos Bitcoin vers le réseau RSK.",
